Rename containing folder.
…and update Storybook index.
