+++ packages/components/src/index.ts
@@ -37,4 +37,4 @@ export {
 	default as HighlightCard,
 	percentCalculator as PercentCalculator,
 } from './highlight-cards/highlight-card';
-export { default as JetpackMobilePromoCard } from './promo-card';
+export { default as JetpackMobilePromoCard } from './jetpack-mobile-promo-card';
