+++ packages/components/src/jetpack-mobile-promo-card/style.scss
