+++ packages/components/src/jetpack-mobile-promo-card/svg-icons.jsx
